SpringDataJPA/data-jpa/src/main/java/study/datajpa/repository/MemberJpaRepository.java
package study.datajpa.repository;

import org.springframework.stereotype.Repository;
import study.datajpa.domain.Member;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberJpaRepository {

    @PersistenceContext
    private EntityManager em;

    public Member save(Member member) {
        em.persist(member);
        return member;
    }

    public Member find(Long id) {
        Member findMember = em.find(Member.class, id);
        return findMember;
    }

    public void flush() {
        em.flush();
        em.clear();
    }
}
