SpringDataJPA/data-jpa/src/test/java/study/datajpa/repository/MemberJpaRepositoryTest.java
package study.datajpa.repository;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.transaction.annotation.Transactional;
import study.datajpa.domain.Member;

import static org.assertj.core.api.Assertions.assertThat;


@SpringBootTest
@Transactional
@Rollback(false)
class MemberJpaRepositoryTest {

    @Autowired
    MemberJpaRepository memberJpaRepository; // 스프링 빈 컨텍스트에서 관리하는 객체들을 자동으로 주입해주는 것은 Spring의 기능

    @Test
    public void testMember() {
        Member member = new Member("memberA");

        Member savedMember = memberJpaRepository.save(member);
//        memberJpaRepository.flush();
        Member findMember = memberJpaRepository.find(savedMember.getId());

        assertThat(findMember.getId()).isEqualTo(savedMember.getId());
        assertThat(findMember.getUsername()).isEqualTo(savedMember.getUsername());

        assertThat(findMember).isEqualTo(member);
        // JPA는 엔티티의 동일성을 보장해준다(1차 캐시). 하지만 member를 save 한 후 flush, clear 하고 find 멤버를 한 경우에는 동일성 보장 X


    }
}